deal with class names and Run + Lumi IOVs
The DB and the EventSetup were not using the same string name to
for the same underlying C++ class. Chose the EventSetup name as
the canonical name.
A Record in the DB used both the Run and Lumi based IOVs for some
of its tags. I now convert Run to Lumi based values early to allow
proper comparisons.